+++ src/gitingest/cli.py
@@ -8,7 +8,7 @@ from typing import Optional, Tuple
 import click
 
 from gitingest.config import MAX_FILE_SIZE, OUTPUT_FILE_NAME
-from gitingest.repository_ingest import ingest_async
+from gitingest.entrypoint import ingest_async
 
 
 @click.command()
